requirements.txt
# Storm-Logos API Requirements

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Auth
pyjwt>=2.8.0

# Database
psycopg2-binary>=2.9.9
neo4j>=5.17.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# LLM Clients
anthropic>=0.18.0
groq>=0.4.0

# NLP
spacy>=3.7.0
numpy>=1.26.0
scipy>=1.11.0

# Data
pydantic>=2.5.0

# Utils
python-dotenv>=1.0.0
requests>=2.31.0
